+++ docs/reference/host-core-reset.md
@@ -88,6 +88,8 @@ The current native-host Swift split is:
   session controller.
 - `LiveOverlayRenderer.swift`: live overlay layer composition for HUD, loupe, scrims, and selection
   affordances.
+- `LiveChromePlacement.swift`: live HUD/loupe text metrics, pending color text, and deterministic
+  floating placement geometry shared by capture host and live chrome rendering.
 - `LiveOverlayWindowSnapshotFeed.swift`, `LiveOverlayChromeSampleFeed.swift`, and
   `LiveFrameClockDriver.swift`: native live overlay support boundaries for target-window snapshots,
   chrome color/patch sampling, and display-rate frame ticks.
